Adicionado as autorizações em authorcontroller

+++ atividade_08/app/Http/Controllers/AuthorController.php
@@ -7,22 +7,21 @@ use Illuminate\Http\Request;
 
 class AuthorController extends Controller
 {
-    // Exibe uma lista de autores
     public function index()
     {
-        $authors = Author::all();
-        return view('authors.index', compact('authors'));
+        return view('authors.index', ['authors' => Author::all()]);
     }
 
-    // Mostra o formulário para criar um novo autor
     public function create()
     {
+        $this->authorize('create', Author::class);
         return view('authors.create');
     }
 
-    // Armazena um novo autor no banco de dados
     public function store(Request $request)
     {
+        $this->authorize('create', Author::class);
+
         $request->validate([
             'name' => 'required|string|unique:authors|max:255',
         ]);
@@ -32,21 +31,21 @@ class AuthorController extends Controller
         return redirect()->route('authors.index')->with('success', 'Autor criado com sucesso.');
     }
 
-    // Exibe um autor específico
     public function show(Author $author)
     {
         return view('authors.show', compact('author'));
     }
 
-    // Mostra o formulário para editar um autor existente
     public function edit(Author $author)
     {
+        $this->authorize('update', $author);
         return view('authors.edit', compact('author'));
     }
 
-    // Atualiza um autor no banco de dados
     public function update(Request $request, Author $author)
     {
+        $this->authorize('update', $author);
+
         $request->validate([
             'name' => 'required|string|unique:authors,name,' . $author->id . '|max:255',
         ]);
@@ -56,9 +55,9 @@ class AuthorController extends Controller
         return redirect()->route('authors.index')->with('success', 'Autor atualizado com sucesso.');
     }
 
-    // Remove um autor do banco de dados
     public function destroy(Author $author)
     {
+        $this->authorize('delete', $author);
         $author->delete();
 
         return redirect()->route('authors.index')->with('success', 'Autor excluído com sucesso.');
